requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
gunicorn==19.9.0
django==3.1.1
pillow==8.1.2
django-phonenumber-field[phonenumbers]==5.0.0
django-filter==2.4.0
django-bootstrap-form
unicodecsv==0.14.1
#for deployment in Heroku
django-heroku==0.3.1
whitenoise==5.2.0
dj-database-url==0.5.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2017.3